+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Feat: Added optimized images for project
 
 ## [1.1.6] - 2024-08-31
 ### Fixed
@@ -264,7 +266,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Few issues on dark mode and event listener
 
-[unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.6...HEAD
+[Unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.6...HEAD
 [1.1.6]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.5...v1.1.6
 [1.1.5]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.3...v1.1.4